// src/api/models/CreatePassengerRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreatePassengerRequest = {
    firstName?: string;
    lastName?: string;
    dateOfBirth?: string;
    gender?: string;
    passportNumber?: string;
    passportCountry?: string;
    passportExpiryDate?: string;
    nationality?: string;
    email?: string;
    phoneNumber?: string;
};